% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydrc_plot.R
\name{tidydrc_plot}
\alias{tidydrc_plot}
\title{Plot model and data from \code{tidydrc()}}
\usage{
tidydrc_plot(modeldf, ed50 = FALSE, confint = FALSE, ...)
}
\arguments{
\item{modeldf}{A list-column dataframe, generated by \code{tidydrc()}}

\item{ed50}{Logical, whether to draw a line indicating the ED50 (effective dose for 50\% response).}

\item{confint}{Logical, whether to draw confidence bands around the fitted regression curve as a \code{geom_ribbon}}

\item{...}{Arguments passed on to \code{aes()} of ggplot, as in \code{ggplot(aes(...))}}
}
\description{
A \code{ggplot} wrapper function for plotting the dose-response model fit and the data from the list-column output of \code{tidydrc()}
}
\examples{
# directly piping the output of tidydrc()
library(tidyverse)
p <- tidydrc(S.alba, Dose, DryMatter, model = LL.4(), Herbicide) \%>\%
     tidydrc_plot(ed50 = TRUE, confint = TRUE, color = ~Herbicide)
p

# the ggplot object can be manipulated further as usual
p + facet_wrap(~Herbicide, ncol = 1) +
    scale_x_log10() +
    labs(title = "Potency of glyphosate and bentazone in Sinapsis alba")

}
\author{
Angel Angelov
}
